Updated dependencies and fixing accept script
