+++ package.json
@@ -33,15 +33,15 @@
   },
   "dependencies": {
     "@agoric/access-token": "^0.4.20",
-    "@agoric/assert": "^0.5.1",
+    "@agoric/assert": "^0.6.0",
     "@agoric/cache": "^0.2.3",
     "@agoric/casting": "^0.3.2",
     "@agoric/cosmic-proto": "^0.2.1",
     "@agoric/ertp": "^0.15.3",
-    "@agoric/inter-protocol": "^0.13.1",
+    "@agoric/inter-protocol": "^0.14.0",
     "@agoric/internal": "^0.2.1",
     "@agoric/smart-wallet": "^0.4.2",
-    "@agoric/store": "^0.8.3",
+    "@agoric/store": "^0.9.0",
     "@agoric/swingset-vat": "^0.30.2",
     "@agoric/vats": "^0.13.0",
     "@agoric/zoe": "^0.25.3",
@@ -62,7 +62,7 @@
     "anylogger": "^0.21.0",
     "body-parser": "^1.20.1",
     "chalk": "^2.4.2",
-    "commander": "^5.0.0",
+    "commander": "^10.0.0",
     "db-migrate": "^0.11.13",
     "dd-trace": "^3.3.0",
     "deterministic-json": "^1.0.5",
