+++ scripts/accept-oracle-invitation.sh
@@ -12,9 +12,8 @@ WALLET_ADDR=$(agd keys show "$WALLET_NAME" --keyring-backend test --output json 
 ORACLE_OFFER=$(mktemp -t agops.XXX)
 ./src/bin-agops.js oracle accept --pair "$BRAND_IN.$BRAND_OUT" >|"$ORACLE_OFFER"
 cat $ORACLE_OFFER
-jq ".body | fromjson" <"$ORACLE_OFFER"
 OFFER_TO_SEND=$(cat $ORACLE_OFFER | jq .)
 agd --chain-id=$CHAIN_ID --keyring-backend=test --from=$WALLET_ADDR tx swingset wallet-action --allow-spend "$OFFER_TO_SEND" --yes --node $AGORIC_RPC
 sleep 3
 
-echo "Accepted oracle invitation"
+echo "Accepted oracle invitation from $WALLET_ADDR"
